Show and modify signal detail assignee
