+++ src/signals/incident-management/containers/IncidentDetail/components/ChangeValue/index.js
@@ -1,0 +1,3 @@
+import ChangeValue from './ChangeValue';
+
+export default ChangeValue;
